Move digamma to new library
